Clean up Router & Add documentation
