+++ app/src/main/java/mozilla/org/webmaker/router/helper/RouterParams.java
@@ -1,0 +1,8 @@
+package mozilla.org.webmaker.router.helper;
+
+import java.util.Map;
+
+public class RouterParams {
+    public RouterOptions routerOptions;
+    public Map<String, String> openParams;
+}
